UseEffect部分のリファクタリング
